requirements.txt
tensorflow>=2.13,<3.0
numpy>=1.24
pillow>=10.0
matplotlib>=3.7
streamlit>=1.30
kaggle>=1.6

